backend/requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests>=2.31.0

# 音频处理 (基础)
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# 语音识别 (简化)
faster-whisper>=1.0.1
# 避免av包编译问题，使用预编译版本
av>=11.0.0; platform_system != "Windows"
# Windows用户可选择跳过av依赖

# 繁简转换
opencc-python-reimplemented==0.1.7

# 文字转语音
edge-tts==6.1.9
pyttsx3>=2.90

# 翻译
translators==5.9.9

# 文档处理
python-docx==1.1.0
openpyxl>=3.1.0

# 图像处理
Pillow==10.1.0

# 其他工具
aiohttp>=3.8.0
websockets>=11.0.0
jieba>=0.42.1

# 桌面应用
pywebview==4.4.1

# 打包
pyinstaller==6.3.0
